+++ ams/__init__.py
@@ -7,6 +7,9 @@ from ams import io  # NOQA
 from ams import utils  # NOQA
 from ams import models     # NOQA
 from ams import system    # NOQA
+from ams import routines  # NOQA
+from ams import opt       # NOQA
+from ams import solver  # NOQA
 
 __author__ = 'Jining Wang'