+++ ams/opt/__init__.py
@@ -1,0 +1,5 @@
+"""
+Module for optimization modeling.
+"""
+
+from ams.opt.omodel import Constraint, Objective, OModel