+++ ams/routines/pflow.py
@@ -1,6 +1,8 @@
 """
 Power flow routines.
 """
+import logging
+from collections import OrderedDict
 
 import numpy as np
 
@@ -11,50 +13,70 @@ from ams.solver.pypower.runpf import runpf, rundcpf
 
 from ams.io.pypower import system2ppc
 
+from ams.utils import timer
+
+logger = logging.getLogger(__name__)
+
 
 class PFlow(BaseRoutine):
     """
-    AC Power flow routine.
+    AC Power Flow routine.
     """
 
     def __init__(self, system=None, config=None):
         super().__init__(system, config)
+        # FIXME: temp solution, adapt to new routine later on
+        self.rparams = OrderedDict()
+        self.ralgebs = OrderedDict()
+        # --- remove above two arrtributes later on ---
         self.info = "AC Power flow"
-        self.converged = False
+        self.rtn_models = OrderedDict([
+            ('Bus', ['vmax', 'vmin']),
+            ('Slack', ['pmax', 'pmin', 'qmax', 'qmin']),
+            ('PV', ['pmax', 'pmin', 'qmax', 'qmin']),
+        ])
+
+    @timer
+    def solve(self, **kwargs):
+        ppc = system2ppc(self.system)
+        res, success = runpf(ppc, **kwargs)
+        return ppc, success
+    
+    def unpack(self, ppc):
+        """
+        Unpack results from ppc to system.
+        """
+        system = self.system
+
+        # --- Bus ---
+        system.Bus.v.v = ppc['bus'][:, 7]  # voltage magnitude
+        system.Bus.a.v = ppc['bus'][:, 8] * deg2rad  # voltage angle
+
+        # --- PV ---
+        system.PV.q.v = ppc['gen'][system.Slack.n:, 2]  # reactive power
+
+        # --- Slack ---
+        system.Slack.p.v = ppc['gen'][:system.Slack.n, 1]  # active power
+        system.Slack.q.v = ppc['gen'][:system.Slack.n, 2]  # reactive power
+
+        # --- store results into routine algeb ---
+        for raname, oalgeb in self.oalgebs.items():
+            oalgeb.v = oalgeb.Algeb.v.copy()
 
     def run(self, **kwargs):
         """
         Run power flow.
         """
-        ppc = system2ppc(self.system)
-        res, exit_code = runpf(ppc, **kwargs)
-        self.converged = bool(exit_code)
-
-        # --- Update variables ---
-        system = self.system
-
-        # --- Bus ---
-        system.Bus.v.v = a_Bus = ppc['bus'][:, 7]  # voltage magnitude
-        system.Bus.a.v = v_Bus = ppc['bus'][:, 8] * deg2rad # voltage angle
-
-        # --- PV ---
-        system.PV.q.v = q_PV = ppc['gen'][system.Slack.n:, 2]  # reactive power
-
-        # --- Slack ---
-        system.Slack.p.v = p_Slack = ppc['gen'][:system.Slack.n, 1]  # active power
-        system.Slack.q.v = q_Slack = ppc['gen'][:system.Slack.n, 2]  # reactive power
-
-        # --- routine ---
-        for aname in self.algebs.keys():
-            a = self.algebs[aname]['a']  # array index
-            if len(a) > 1:
-                self.v[a[0]:a[-1]+1] = locals()[aname]
-                self.algebs[aname]['algeb'].v = locals()[aname]
-            else:
-                self.v[a] = locals()[aname]
-                self.algebs[aname]['algeb'].v = locals()[aname]
-
-        return self.converged
+        (ppc, success), elapsed_time = self.solve(**kwargs)
+        self.exit_code = int(not success)
+        if success:
+            info = f'{self.class_name} completed in {elapsed_time} seconds with exit code {self.exit_code}.'
+        else:
+            info = f'{self.class_name} failed in {elapsed_time} seconds with exit code {self.exit_code}.'
+        logger.info(info)
+        self.exec_time = float(elapsed_time.split(' ')[0])
+        self.unpack(ppc)
+        return self.exit_code
 
     def summary(self, **kwargs):
         """
@@ -69,12 +91,14 @@ class PFlow(BaseRoutine):
         pass
 
 
-class DCPF(BaseRoutine):
+class DCPF(PFlow):
     """
-    DC Power flow routine.
+    DC Power Flow routine.
     """
+
     def __init__(self, system=None, config=None):
         super().__init__(system, config)
+        self.info = "DC Power Flow"
 
     def run(self, **kwargs):
         """