+++ docs/source/release-notes.rst
@@ -8,11 +8,22 @@ The APIs before v3.0.0 are in beta and may change without prior notice.
 
 Pre-v1.0.0
 ==========
-v0.6.2
+
+v0.6.3 (2023-05-22)
+-------------------
+
+- Using CVXPY for draft implementation
+- Improve ``model``, ``group``, ``param`` and ``var`` in ``core``
+- Refactor ``routines`` and ``opt``
+- Improve PYPOWER interface ``io.pypower.system2ppc``
+- Fix PYPOWER function ``solver.pypower.makePTDF``
+
+v0.6.2 (2023-04-23)
 -------------------
 
 - Enhance docstring
 - Remove unused module ``utils.LazyImport``
+- Remove unused module ``shared``
 
 v0.6.1 (2023-03-05)
 -------------------
@@ -21,7 +32,7 @@ v0.6.1 (2023-03-05)
 - Add file parser ``io.pypower`` for PYPOWER case file
 - Deprecate PYPOWER interface ``solvers.ipp``
 
-v0.6 (2023-03-04)
+v0.6.0 (2023-03-04)
 -------------------
 
 - Set up PYPOWER for power flow calculation
@@ -35,8 +46,8 @@ v0.6 (2023-03-04)
 v0.5 (2023-02-17)
 -------------------
 
-- Base System setup
-- Development preparation
+- Develop module ``system``, ``main``, ``cli``
+- Development preparation: versioneer, documentation, etc.
 
 v0.4 (2023-01)
 -------------------